+++ alpha/external-web/app/routes.js
@@ -4,9 +4,9 @@ var router = express.Router()
 module.exports = router
 
 // Include custom route files.
-require('./index')
-require('./about-you')
-require('./relationship')
-require('./escorts')
-require('./about-your-income')
-require('./application-submitted')
+require('./routes/index')
+require('./routes/about-you')
+require('./routes/relationship')
+require('./routes/escorts')
+require('./routes/about-your-income')
+require('./routes/application-submitted')
