+++ alpha/external-web/app/routes/about-you.js
@@ -1,10 +1,10 @@
 /**
  * This file defines all routes for the about-you page.
  */
-var router = require('./routes')
+var router = require('../routes')
 
 // TODO: Should be included in a controller file rather than by each routes file.
-var mongo = require('./database')
+var mongo = require('../database')
 
 /**
  * Save a single claimant to the system.
