+++ alpha/external-web/app/routes/about-your-income.js
@@ -1,11 +1,15 @@
 /**
  * This file defines all routes for the about-your-income page.
  */
-var router = require('./routes')
+var router = require('../routes')
 
+// TODO: Should be included bu a controller file.
 // Require file upload library.
 var multer = require('multer')
 var upload = multer({ dest: 'eligibility-uploads/' })
+
+// TODO: Should be included in a controller file rather than by each routes file.
+var mongo = require('../database')
 
 /**
  * Render the about you income page.
