+++ alpha/external-web/app/routes/application-submitted.js
@@ -1,7 +1,7 @@
 /**
  * This file defines all routes for the application-submitted page.
  */
-var router = require('./routes')
+var router = require('../routes')
 
 /**
  * Render the application submitted page. Displayed after a successful file upload.
