+++ alpha/external-web/app/routes/index.js
@@ -1,7 +1,7 @@
 /**
  * This file defines all routes for the index page.
  */
-var router = require('./routes')
+var router = require('../routes')
 
 /**
  * Render the landing page.
